+++ package.json
@@ -51,6 +51,7 @@
     "eslint-plugin-jsdoc": "^39.6.7",
     "eslint-plugin-sf-plugin": "^1.9.2",
     "husky": "^7.0.4",
+    "jsforce": "^2.0.0-beta.20",
     "mocha": "^10.1.0",
     "nyc": "^15.1.0",
     "oclif": "^3.7.0",
